:memo: persist 폴더 구조로 refactor
