+++ user/service/src/main/kotlin/com/gongsung/user/UserService.kt
@@ -1,7 +1,7 @@
 package com.gongsung.user
 
 import com.gongsung.user.command.CommandUserUseCase
-import com.gongsung.user.persist.CommandUserPersist
+import com.gongsung.user.persist.user.CommandUserPersist
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
